fix: merge deviceObserver and useDeviceData hooks
