+++ packages/framework/src/hooks/index.ts
@@ -13,3 +13,4 @@ export * from "./useThemeContext";
 export * from "./useEventContext";
 export * from "./useLanguageScope";
 export * from "./useScreenData";
+export * from "./useDeviceObserver";
