+++ packages/framework/src/hooks/useScreenContext.tsx
@@ -1,6 +1,6 @@
 import { Provider, useAtom, useAtomValue, useSetAtom } from "jotai";
-import { useCallback, useContext, useEffect, useMemo } from "react";
-import { clone, debounce, merge } from "lodash-es";
+import { useCallback, useContext } from "react";
+import { clone, merge } from "lodash-es";
 import { useHydrateAtoms } from "jotai/utils";
 import {
   appAtom,
@@ -10,7 +10,6 @@ import {
   screenDataAtom,
   themeAtom,
   userAtom,
-  useDeviceData,
 } from "../state";
 import type {
   ApplicationContextDefinition,
@@ -21,6 +20,7 @@ import type { Response, WebSocketConnection } from "../data";
 import type { EnsembleScreenModel } from "../shared/models";
 import { useApplicationContext } from "./useApplicationContext";
 import { CustomThemeContext } from "./useThemeContext";
+import { useDeviceObserver } from "./useDeviceObserver";
 
 interface ScreenContextProps {
   screen: EnsembleScreenModel;
@@ -118,28 +118,4 @@ export const useScreenContext = ():
   );
 
   return { ...screenContext, setData };
-};
-
-const useDeviceObserver = (): void => {
-  const { setData: updateDeviceData } = useDeviceData();
-
-  const debouncedUpdateDeviceData = useMemo(() => {
-    const handleResize = (): void => {
-      updateDeviceData({
-        width: window.innerWidth,
-        height: window.innerHeight,
-      });
-    };
-    return debounce(handleResize, 100);
-  }, [updateDeviceData]);
-
-  useEffect(() => {
-    const resizeObserver = new ResizeObserver(debouncedUpdateDeviceData);
-    resizeObserver.observe(document.body);
-
-    return (): void => {
-      resizeObserver.disconnect();
-      debouncedUpdateDeviceData.cancel(); // Cancel any pending debounced calls on cleanup
-    };
-  }, [debouncedUpdateDeviceData]);
 };
