+++ packages/framework/src/state/index.ts
@@ -3,4 +3,3 @@ export * from "./screen";
 export * from "./application";
 export * from "./platform";
 export * from "./user";
-export * from "./device";
