+++ packages/runtime/src/widgets/Form/__tests__/Form.test.tsx
@@ -6,6 +6,9 @@ import type { PropsWithChildren } from "react";
 import { BrowserRouter } from "react-router-dom";
 import { Button, Form, Text } from "../../index";
 
+// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+global.ResizeObserver = require("resize-observer-polyfill");
+
 const FormTestWrapper: React.FC<PropsWithChildren> = ({ children }) => {
   return (
     <BrowserRouter>
